move pkg to kernel
