+++ internal/producer/producer.go
@@ -28,7 +28,6 @@ import (
 	"github.com/apache/rocketmq-client-go/internal/remote"
 	"github.com/apache/rocketmq-client-go/primitive"
 	"github.com/apache/rocketmq-client-go/rlog"
-	"github.com/apache/rocketmq-client-go/utils"
 	"github.com/pkg/errors"
 )
 
@@ -40,7 +39,7 @@ type Producer interface {
 }
 
 func NewProducer(nameServerAddrs []string, opts ...*primitive.ProducerOption) (Producer, error) {
-	srvs, err := utils.NewNamesrv(nameServerAddrs...)
+	srvs, err := kernel.NewNamesrv(nameServerAddrs...)
 	if err != nil {
 		return nil, errors.Wrap(err, "new Namesrv failed.")
 	}
